requirements.txt
fastapi
uvicorn[standard]
tensorflow
pillow
numpy
gdown
python-multipart